requirements.txt
diffusers==0.21.2
einops
gdown
jaxtyping
kornia
opencv-python
pytorch-lightning==2.0.9
taming-transformers-rom1504
tqdm
typeguard
transformers==4.28.1
git+https://github.com/openai/CLIP.git
git+https://github.com/chinhsuanwu/liegroups.git